test: 인수 테스트 통합
내용:
- 시나리오, 흐름을 검증하는 테스트로 리팩터링
- 관련 동작 별도의 Actions 클래스로 분리
